chore: Refactor repository code
