+++ src/SaanSoft.Cqrs.Decorator.Store.MongoDB/CommandRepository.cs
@@ -1,27 +1,25 @@
-using MongoDB.Driver;
-using SaanSoft.Cqrs.Messages;
-
 namespace SaanSoft.Cqrs.Decorator.Store.MongoDB;
 
-public interface ICommandMongoDbRepository<TMessageId> : ICommandRepository<TMessageId>, IMongoDbStore<TMessageId>
+public interface ICommandMongoDbRepository<TMessageId> :
+    ICommandRepository<TMessageId>,
+    IMongoDbRepository
     where TMessageId : struct
+{
+    IMongoCollection<IMessage<TMessageId>> MessageCollection { get; }
+}
+
+public class CommandRepository(IMongoDatabase database, IIdGenerator<Guid> idGenerator)
+    : CommandRepository<Guid>(database, idGenerator)
 {
 }
 
-public class CommandRepository(IMongoDatabase database)
-    : CommandRepository<Guid>(database)
-{
-    protected override Guid NewMessageId() => Guid.NewGuid();
-}
-
-public abstract class CommandRepository<TMessageId>(IMongoDatabase database) :
-    BaseMessageRepository<TMessageId, ICommand<TMessageId>>(database),
-    ICommandMongoDbRepository<TMessageId>,
-    ICommandPublisherRepository<TMessageId>,
-    ICommandHandlerRepository<TMessageId>
+public abstract class CommandRepository<TMessageId>(IMongoDatabase database, IIdGenerator<TMessageId> idGenerator) :
+    BaseMessageRepository<TMessageId, ICommand<TMessageId>>(database, idGenerator),
+    ICommandMongoDbRepository<TMessageId>
     where TMessageId : struct
 {
-    public override string MessageCollectionName => "CommandMessages";
-    public override string PublisherCollectionName => "CommandPublishers";
-    public override string HandlerCollectionName => "CommandHandlers";
+    public override string CollectionName => "CommandMessages";
+
+    public IMongoCollection<IMessage<TMessageId>> MessageCollection
+        => Database.GetCollection<IMessage<TMessageId>>(CollectionName);
 }
