+++ src/SaanSoft.Cqrs.Decorator.Store.MongoDB/GlobalUsings.cs
@@ -1,0 +1,4 @@
+global using MongoDB.Driver;
+global using SaanSoft.Cqrs.Decorator.Store.Models;
+global using SaanSoft.Cqrs.Messages;
+global using SaanSoft.Cqrs.Utilities;
