+++ src/SaanSoft.Cqrs.Decorator.Store.MongoDB/IMongoDBRepository.cs
@@ -1,0 +1,8 @@
+namespace SaanSoft.Cqrs.Decorator.Store.MongoDB;
+
+public interface IMongoDbRepository
+{
+    string CollectionName { get; }
+
+    Task EnsureCollectionIndexes(CancellationToken cancellationToken = default);
+}
