--- src/SaanSoft.Cqrs.Decorator.Store.MongoDB/IMongoDbStore.cs
@@ -1,9 +1,0 @@
-using MongoDB.Driver;
-using SaanSoft.Cqrs.Messages;
-
-namespace SaanSoft.Cqrs.Decorator.Store.MongoDB;
-
-public interface IMongoDbStore<TMessageId> where TMessageId : struct
-{
-    IMongoCollection<IMessage<TMessageId>> MessageCollection { get; }
-}
