+++ src/SaanSoft.Cqrs.Decorator.Store.MongoDB/QueryRepository.cs
@@ -1,26 +1,25 @@
-using MongoDB.Driver;
-using SaanSoft.Cqrs.Messages;
-
 namespace SaanSoft.Cqrs.Decorator.Store.MongoDB;
 
-public interface IQueryMongoDbRepository<TMessageId> : IQueryRepository<TMessageId>, IMongoDbStore<TMessageId>
+public interface IQueryMongoDbRepository<TMessageId> :
+    IQueryRepository<TMessageId>,
+    IMongoDbRepository
     where TMessageId : struct
+{
+    IMongoCollection<IMessage<TMessageId>> MessageCollection { get; }
+}
+
+public class QueryRepository(IMongoDatabase database, IIdGenerator<Guid> idGenerator, InsertManyOptions? insertManyOptions = null) :
+    QueryRepository<Guid>(database, idGenerator, insertManyOptions)
 {
 }
 
-public class QueryRepository(IMongoDatabase database) : QueryRepository<Guid>(database)
-{
-    protected override Guid NewMessageId() => Guid.NewGuid();
-}
-
-public abstract class QueryRepository<TMessageId>(IMongoDatabase database) :
-    BaseMessageRepository<TMessageId, IQuery<TMessageId>>(database),
-    IQueryMongoDbRepository<TMessageId>,
-    IQueryPublisherRepository<TMessageId>,
-    IQueryHandlerRepository<TMessageId>
+public abstract class QueryRepository<TMessageId>(IMongoDatabase database, IIdGenerator<TMessageId> idGenerator, InsertManyOptions? insertManyOptions = null) :
+    BaseMessageRepository<TMessageId, IQuery<TMessageId>>(database, idGenerator, insertManyOptions),
+    IQueryMongoDbRepository<TMessageId>
     where TMessageId : struct
 {
-    public override string MessageCollectionName => "QueryMessages";
-    public override string PublisherCollectionName => "QueryPublishers";
-    public override string HandlerCollectionName => "QueryHandlers";
+    public override string CollectionName => "QueryMessages";
+
+    public IMongoCollection<IMessage<TMessageId>> MessageCollection
+        => Database.GetCollection<IMessage<TMessageId>>(CollectionName);
 }
