+++ src/SaanSoft.Cqrs.Decorator.Store.MongoDB/SaanSoft.Cqrs.Decorator.Store.MongoDB.csproj
@@ -9,7 +9,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\SaanSoft.Cqrs.Decorator.GuidIds\SaanSoft.Cqrs.Decorator.GuidIds.csproj" />
+    <ProjectReference Include="..\SaanSoft.Cqrs.GuidIds\SaanSoft.Cqrs.GuidIds.csproj" />
     <ProjectReference Include="..\SaanSoft.Cqrs.Decorator\SaanSoft.Cqrs.Decorator.csproj" />
     <ProjectReference Include="..\SaanSoft.Cqrs\SaanSoft.Cqrs.csproj" />
   </ItemGroup>
