+++ src/SaanSoft.Cqrs.GuidIds/Bus/InMemoryCommandBus.cs
@@ -1,4 +1,4 @@
-namespace SaanSoft.Cqrs.Decorator.GuidIds.Bus;
+namespace SaanSoft.Cqrs.GuidIds.Bus;
 
 public class InMemoryCommandBus(IServiceProvider serviceProvider, IIdGenerator<Guid> idGenerator, ILogger logger)
     : InMemoryCommandBus<Guid>(serviceProvider, idGenerator, logger);
