+++ src/SaanSoft.Cqrs.GuidIds/Bus/InMemoryEventBus.cs
@@ -1,4 +1,4 @@
-namespace SaanSoft.Cqrs.Decorator.GuidIds.Bus;
+namespace SaanSoft.Cqrs.GuidIds.Bus;
 
 public class InMemoryEventBus(IServiceProvider serviceProvider, IIdGenerator<Guid> idGenerator, ILogger logger)
     : InMemoryEventBus<Guid>(serviceProvider, idGenerator, logger);
