+++ src/SaanSoft.Cqrs.GuidIds/Bus/InMemoryQueryBus.cs
@@ -1,4 +1,4 @@
-namespace SaanSoft.Cqrs.Decorator.GuidIds.Bus;
+namespace SaanSoft.Cqrs.GuidIds.Bus;
 
 public class InMemoryQueryBus(IServiceProvider serviceProvider, IIdGenerator<Guid> idGenerator, ILogger logger)
     : InMemoryQueryBus<Guid>(serviceProvider, idGenerator, logger);
