+++ src/SaanSoft.Cqrs.GuidIds/Decorator/Store/StoreCommandHandlerDecorator.cs
@@ -1,0 +1,6 @@
+using SaanSoft.Cqrs.Decorator.Store;
+
+namespace SaanSoft.Cqrs.GuidIds.Decorator.Store;
+
+public class StoreCommandHandlerDecorator(ICommandHandlerRepository<Guid> repository, ICommandSubscriptionBus<Guid> next)
+    : StoreCommandHandlerDecorator<Guid>(repository, next);
