+++ src/SaanSoft.Cqrs.GuidIds/Decorator/Store/StoreCommandPublisherDecorator.cs
@@ -1,0 +1,6 @@
+using SaanSoft.Cqrs.Decorator.Store;
+
+namespace SaanSoft.Cqrs.GuidIds.Decorator.Store;
+
+public class StoreCommandPublisherDecorator(ICommandPublisherRepository<Guid> repository, ICommandBus<Guid> next)
+    : StoreCommandPublisherDecorator<Guid>(repository, next);
