+++ src/SaanSoft.Cqrs.GuidIds/Decorator/Store/StoreEventHandlerDecorator.cs
@@ -1,0 +1,6 @@
+using SaanSoft.Cqrs.Decorator.Store;
+
+namespace SaanSoft.Cqrs.GuidIds.Decorator.Store;
+
+public class StoreEventHandlerDecorator(IEventHandlerRepository<Guid> repository, IEventSubscriptionBus<Guid> next)
+    : StoreEventHandlerDecorator<Guid>(repository, next);
