+++ src/SaanSoft.Cqrs.GuidIds/Decorator/Store/StoreEventPublisherDecorator.cs
@@ -1,0 +1,6 @@
+using SaanSoft.Cqrs.Decorator.Store;
+
+namespace SaanSoft.Cqrs.GuidIds.Decorator.Store;
+
+public class StoreEventPublisherDecorator(IEventPublisherRepository<Guid> repository, IEventBus<Guid> next)
+    : StoreEventPublisherDecorator<Guid>(repository, next);
