+++ src/SaanSoft.Cqrs.GuidIds/Decorator/Store/StoreQueryHandlerDecorator.cs
@@ -1,0 +1,6 @@
+using SaanSoft.Cqrs.Decorator.Store;
+
+namespace SaanSoft.Cqrs.GuidIds.Decorator.Store;
+
+public class StoreQueryHandlerDecorator(IQueryHandlerRepository<Guid> repository, IQuerySubscriptionBus<Guid> next)
+    : StoreQueryHandlerDecorator<Guid>(repository, next);
