+++ src/SaanSoft.Cqrs.GuidIds/Decorator/Store/StoreQueryPublisherDecorator.cs
@@ -1,0 +1,6 @@
+using SaanSoft.Cqrs.Decorator.Store;
+
+namespace SaanSoft.Cqrs.GuidIds.Decorator.Store;
+
+public class StoreQueryPublisherDecorator(IQueryPublisherRepository<Guid> repository, IQueryBus<Guid> next)
+    : StoreQueryPublisherDecorator<Guid>(repository, next);
