+++ test/SaanSoft.Tests.Cqrs.Common/GlobalUsings.cs
@@ -2,8 +2,8 @@ global using FakeItEasy;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.Logging;
 global using SaanSoft.Cqrs.Bus;
-global using SaanSoft.Cqrs.Decorator.GuidIds.Bus;
-global using SaanSoft.Cqrs.Decorator.GuidIds.Messages;
+global using SaanSoft.Cqrs.GuidIds.Bus;
+global using SaanSoft.Cqrs.GuidIds.Messages;
 global using SaanSoft.Cqrs.Handler;
 global using SaanSoft.Cqrs.Messages;
 global using SaanSoft.Tests.Cqrs.Common.TestModels;
