+++ test/SaanSoft.Tests.Cqrs.Common/SaanSoft.Tests.Cqrs.Common.csproj
@@ -20,7 +20,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\..\src\SaanSoft.Cqrs.Decorator.GuidIds\SaanSoft.Cqrs.Decorator.GuidIds.csproj" />
+      <ProjectReference Include="..\..\src\SaanSoft.Cqrs.GuidIds\SaanSoft.Cqrs.GuidIds.csproj" />
       <ProjectReference Include="..\..\src\SaanSoft.Cqrs\SaanSoft.Cqrs.csproj" />
     </ItemGroup>
