+++ test/SaanSoft.Tests.Cqrs.Decorator.Store.MongoDB/GlobalUsings.cs
@@ -2,8 +2,8 @@ global using AutoFixture.Xunit2;
 global using FluentAssertions;
 global using Microsoft.Extensions.DependencyInjection;
 global using MongoDB.Driver;
-global using SaanSoft.Cqrs.Decorator.GuidIds.Messages;
 global using SaanSoft.Cqrs.Decorator.Store.Models;
 global using SaanSoft.Cqrs.Decorator.Store.MongoDB;
+global using SaanSoft.Cqrs.GuidIds.Messages;
 global using SaanSoft.Cqrs.Messages;
 global using SaanSoft.Tests.Cqrs.Common.TestModels;
