+++ test/SaanSoft.Tests.Cqrs.Decorator/GlobalUsings.cs
@@ -4,6 +4,7 @@ global using Microsoft.Extensions.DependencyInjection;
 global using SaanSoft.Cqrs.Bus;
 global using SaanSoft.Cqrs.Decorator;
 global using SaanSoft.Cqrs.Decorator.Store;
+global using SaanSoft.Cqrs.GuidIds.Decorator.Store;
 global using SaanSoft.Cqrs.Handler;
 global using SaanSoft.Cqrs.Messages;
 global using SaanSoft.Tests.Cqrs.Common;
