+++ test/SaanSoft.Tests.Cqrs/GlobalUsings.cs
@@ -2,9 +2,8 @@ global using AutoFixture.Xunit2;
 global using FakeItEasy;
 global using FluentAssertions;
 global using Microsoft.Extensions.DependencyInjection;
-global using SaanSoft.Cqrs.Decorator.GuidIds.Bus;
+global using SaanSoft.Cqrs.GuidIds.Bus;
 global using SaanSoft.Cqrs.Handler;
-global using SaanSoft.Cqrs.Utilities;
 global using SaanSoft.Tests.Cqrs.Common;
 global using SaanSoft.Tests.Cqrs.Common.AutoFixture;
 global using SaanSoft.Tests.Cqrs.Common.TestModels;
